Create README.md
Dento is modest system thats I decided to practice with for Web Programming course (II Software Engineering Student).  From the technical part it has the main generalities that a system should be based on, from the viewers side, it is as well 'readable' and has a simple UI/UX design.